titanic_with_api/requirements.txt
alembic==1.7.3
atomicwrites==1.4.0
attrs==21.2.0
autopage==0.4.0
backports.functools-lru-cache==1.6.4
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.0
click==8.0.1
cliff==3.9.0
cloudpickle==2.0.0
cmaes==0.8.2
cmd2==2.2.0
colorama==0.4.4
colorlog==6.4.1
cryptography==3.4.7
cycler==0.10.0
et-xmlfile==1.1.0
Flask==2.0.1
future==0.18.2
greenlet==1.1.2
hpsklearn==0.0.3
hyperopt==0.2.5
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.2
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib==3.4.3
more-itertools==8.10.0
networkx==2.6.3
nose==1.3.7
numpy==1.21.2
olefile==0.46
openpyxl==3.0.9
optuna==2.9.1
packaging==21.0
pandas==1.3.3
patsy==0.5.2
pbr==5.6.0
Pillow==8.3.2
pip==21.2.4
pluggy==1.0.0
prettytable==2.2.1
py==1.10.0
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyreadline==2.1
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
setuptools==58.0.4
six==1.16.0
SQLAlchemy==1.4.25
statsmodels==0.12.2
stevedore==3.4.0
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.1
wheel==0.37.0
win-inet-pton==1.1.0
zipp==3.5.1
